user

+++ Hackathon_users/views.py
@@ -7,7 +7,7 @@ from rest_framework import status
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework.exceptions import PermissionDenied
 from Hackathon_users.models import CustomUser
-from .serializers import EmailVerificationSerializer, LogOutSerializer, LogInSerializer, OtpVerifySerializer, RegisterSerializer, RegisterSerializer2
+from .serializers import EmailVerificationSerializer, LogOutSerializer, LogInSerializer, OtpRequestSerializer, OtpVerifySerializer, RegisterSerializer, RegisterSerializer2
 from rest_framework_simplejwt.exceptions import TokenError
 from django.contrib.auth.signals import user_logged_in, user_logged_out
 from .permissions import *
@@ -122,10 +122,15 @@ def email_link(request, user_data):
     Util.send_email(data)
     return Response(user_data, status=status.HTTP_201_CREATED)
 
-def otp(serializer):
-    phone = serializer.validated_data['phone'] 
-    sms = MessageHandler(f'+234{phone}', serializer.validated_data['phone_otp'])
-    sms.send_otp_to_phone() 
+def otp(phone, otp):
+    if phone[0] == '0':
+        phone_num = phone[1:]
+        sms = MessageHandler(f'+234{phone_num}', otp)
+        sms.send_otp_to_phone() 
+    
+    else:
+        sms = MessageHandler(f'+234{phone}', otp)
+        sms.send_otp_to_phone() 
 
 class RegisterView(APIView):
 
@@ -136,50 +141,55 @@ class RegisterView(APIView):
     @action(methods=['POST'], detail=True)
     def post(self, request):
         user_data = request.data
-        user_data['phone_otp'] = generateOTP()
 
-        serializer = self.serializer_class(data=user_data)
-        if serializer.is_valid():
-            serializer.save()
-            email_link(request=request, user_data=user_data)
-            data={
-                'message': 'success',
-                'verify': 'verify email'
-            }
+        if eval(user_data['age']) >= 18:
+            serializer = self.serializer_class(data=user_data)
+            if serializer.is_valid():
+                serializer.save()
+                email_link(request=request, user_data=user_data)
+                data={
+                    'message': 'success',
+                    'verify': 'verify email'
+                }
 
-            return Response(data, status=status.HTTP_201_CREATED)
+                return Response(data, status=status.HTTP_201_CREATED)
 
+            
         
-    
-        else:
-            obj = CustomUser.objects.get(phone=user_data['phone'])
-        
-            if obj.is_otp_verified == False or obj.is_email_verified == False:
-                obj.delete()
+            else:
+                obj = CustomUser.objects.get(email=user_data['email'])
+            
+                if obj.is_otp_verified == False or obj.is_email_verified == False:
+                    obj.delete()
 
-                serializer1 = self.serializer_class(data=user_data)
-                if serializer1.is_valid():  
-                    serializer1.save()
-                    email_link(request=request, user_data=user_data)
-                    data={
-                        'message': 'success',
-                        'next_page': 'verify email'
-                    }
+                    serializer1 = self.serializer_class(data=user_data)
+                    if serializer1.is_valid():  
+                        serializer1.save()
+                        email_link(request=request, user_data=user_data)
+                        data={
+                            'message': 'success',
+                            'next_page': 'verify email'
+                        }
 
-                    return Response(data, status=status.HTTP_201_CREATED)
+                        return Response(data, status=status.HTTP_201_CREATED)
+
+                    else:
+                        data = {
+                            'error': 'an error occured try again'
+                        }
+
+                        return Response(data, status=status.HTTP_400_BAD_REQUEST)
 
                 else:
                     data = {
-                        'error': 'an error occured try again'
+                        'message': 'failed to create. user already exists',
                     }
-
-                    return Response(data, status=status.HTTP_400_BAD_REQUEST)
-
-            else:
-                data = {
-                    'message': 'failed to create. user already exists',
-                }
                 return Response(data, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            data = {
+                'message': 'failed to create. People below 18 cannot be allowed to register',
+            }
+            return Response(data, status=status.HTTP_400_BAD_REQUEST)
 
 class VerifyEmail(views.APIView):
     serializer_class = EmailVerificationSerializer
@@ -206,7 +216,8 @@ class VerifyEmail(views.APIView):
 @api_view(["POST"])
 def verify_otp(request):
     if request.method == "POST":
-        try:
+        phone=request.data['phone']
+        try:    
             obj = CustomUser.objects.get(phone=request.data['phone'])
             if obj:
                 if obj.is_email_verified == True:
@@ -231,4 +242,37 @@ def verify_otp(request):
                     return Response(data={'message': 'email has not been verified'}, status=status.HTTP_400_BAD_REQUEST)
         except CustomUser.DoesNotExist:
             raise NotFound(detail={'message': 'Permission denied. Phone number not found in the database'})
+
+
+def email_verif(email):
+    try:
+        email = CustomUser.objects.get(email=email)
+        if email.is_email_verified == True:
+            return email
+    
+    except:
+        raise PermissionDenied(detail={'message': 'email has not been verified or does not exist'})
+
+@swagger_auto_schema(method="post",request_body=OtpRequestSerializer())
+@api_view(["POST"])
+def request_otp(request):
+    data = request.data
+    phone = data['phone']   
+    email=data['email']
+ 
+    email2 = email_verif(email)
+    try:
+        phone_check = CustomUser.objects.filter(phone=phone)
+        if phone_check != []:
+            email2.phone = phone
+            email2.phone_otp = generateOTP()
+            email2.save()
+            otp(email2.phone, email2.phone_otp)
+            return Response(data={'message': 'phone otp sent'})            
+        else:
+            raise PermissionDenied(detail={'message': 'this number is already ien use by another account'})
+            
+    except CustomUser.DoesNotExist:
+            raise PermissionDenied(detail={'message': 'this number ris already in use by another account'})
+
